engine: extract esm/layers/orbit.js
